refactor: move the random color to custom schemes file
